Add component installer SPI
